Have a __main__.py and still a cli.py
