+++ setup.cfg
@@ -28,7 +28,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    vault = vault_cli.vault:main
+    vault = vault_cli.cli:main
 
 [options.extras_require]
 hvac =
